==> code/1_aquisicao_e_tratamento_de_dados/1_obtendo_segmet_setor_acoes.ipynb <==
Instalando Bilbiotecas

# ! pip install selenium
# ! pip install webdriver_manager

Importando bibliotecas e instalando o webdriver atualizado

from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.common.by import By
from selenium.webdriver.support.wait import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC

service = Service(ChromeDriverManager().install())

Definindo a url alvo

# # fonte setores
# url = "https://www.b3.com.br/pt_br/produtos-e-servicos/negociacao/renda-variavel/acoes/consultas/classificacao-setorial/"
# # fonte tickers 
# url = "https://www.dadosdemercado.com.br/bolsa/acoes"

Automatizando o download das listas

# abrindo o navegador
driver = webdriver.Chrome(service=None)

# acessando a url
driver.get(url)

# esperando pelo botão de aceitar coockies e clicando nele
WebDriverWait(driver, 5).until(
    EC.visibility_of_element_located(
        (By.XPATH, "//button[@id='onetrust-accept-btn-handler']")
        )).click()

# esperando o carregamento do iframe e mudando o 
# induza WebDriverWait para que o elemento desejado seja clicável
# WebDriverWait(driver, 5).until(
#     EC.frame_to_be_available_and_switch_to_it(
#         (By.XPATH,"//iframe[@id='bvmf_iframe']")))

WebDriverWait(driver, 5).until(
    EC.frame_to_be_available_and_switch_to_it(
        (By.XPATH,"/html/body/app-root/app-companies-classification/div']")))


# esperando pela caixa de seleção do número de itens exibidos na lista e clicando nela
WebDriverWait(driver, 5).until(
    EC.visibility_of_element_located(
        (By.XPATH,"/html/body/app-root/app-companies-classification/div/div/div/app-companies-home-filter-classification/form/div[1]/div/div/select")
        )).click()



# Fechando o navegador
# driver.quit()


# # o elemento TODOS está dentro de um <iframe>, então:
# # induza o WebDriver, espere até que o frame desejado esteja disponível e mude para ele;
# # induza WebDriverWait para que o elemento desejado seja clicável
# WebDriverWait(driver, 5).until(
#     EC.frame_to_be_available_and_switch_to_it(
#         (By.XPATH,"//iframe[@id='bvmf_iframe']")))

# # esperando pelo botão TODOS e clicando nele
# WebDriverWait(driver, 5).until(
#     EC.visibility_of_element_located(
#         (By.XPATH, "//button[@aria-label='Buscar Todas']"))).click()

# # esperando pelo botão exibir como lista e clicando nele
# WebDriverWait(driver, 5).until(
#     EC.visibility_of_element_located(
#         (By.XPATH, "/html/body/app-root/app-companies-search/div/form/div[1]/div/div/div[2]/nav/div/a[2]")
#         )).click()

# # esperando pela caixa de seleção do número de itens exibidos na lista e clicando nela
# WebDriverWait(driver, 5).until(
#     EC.visibility_of_element_located(
#         (By.XPATH,"/html/body/app-root/app-companies-search/div/form/div[3]/div[1]/select")
#         )).click()

# # esperando pela abertura das opções da caixa de seleção do número 
# # de itens exibidos na lista e clicando na opção 120
# WebDriverWait(driver, 5).until(
#     EC.visibility_of_element_located(
#         (By.XPATH,"/html/body/app-root/app-companies-search/div/form/div[3]/div[1]/select/option[4]")
#         )).click()

# # Espere até que algum elemento esteja visível ou alguma condição seja atendida
# empresas = WebDriverWait(driver, 5).until(
#     EC.presence_of_element_located(
#         (By.XPATH,'/html/body/app-root/app-companies-search/div/form/div[2]/div[2]/table'))
#         ).text
