==> code/1_aquisicao_e_tratamento_de_dados/2_tratando_segmet_setor_acoes.ipynb <==
Instalando as bibliotecas

# ! pip install pip install pandas
# ! pip install openpyxl

Importando as bibliotecas

import pandas as pd
from openpyxl import Workbook, load_workbook

Estruturando dados em um df

lista_setores = []
lista_subsetores = []
lista_nomes = []
lista_codigos = []

arquivo = f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\empresas_por_setor_B3_01-11-2023.xlsx"

planilha = load_workbook(arquivo)

aba_ativa = planilha.active

for celula in aba_ativa["A"]:
    lista_setores.append(celula.value)

for celula in aba_ativa["B"]:
    lista_subsetores.append(celula.value)

for celula in aba_ativa["C"]:
    lista_nomes.append(celula.value)

for celula in aba_ativa["D"]:
    lista_codigos.append(celula.value)

df_empresas_b3 = pd.DataFrame({"nome": lista_nomes,
                               "codigo": lista_codigos, 
                               "setor": lista_setores,
                               "subsetor": lista_subsetores})

Tratando os dados

df_empresas_b3.dropna(axis=0, how='all', inplace=True)

df_empresas_b3 = df_empresas_b3.loc[
    df_empresas_b3["setor"] != "CLASSIFICAÇÃO SETORIAL DAS EMPRESAS NEGOCIADAS NA B3"]

df_empresas_b3 = df_empresas_b3.loc[df_empresas_b3["setor"] != "SETOR ECONÔMICO"]

df_empresas_b3 = df_empresas_b3.loc[df_empresas_b3["codigo"] != "CÓDIGO"]

df_empresas_b3["segmento"] = df_empresas_b3.loc[df_empresas_b3["codigo"].isna(), "nome"]

df_empresas_b3["setor"] = df_empresas_b3['setor'].ffill()

df_empresas_b3["subsetor"] = df_empresas_b3['subsetor'].ffill()

df_empresas_b3["segmento"] = df_empresas_b3['segmento'].ffill()

df_empresas_b3 = df_empresas_b3.loc[df_empresas_b3["codigo"].notna()]

df_empresas_b3['nome'] = df_empresas_b3['nome'].str.strip()

df_empresas_b3.reset_index(drop=True, inplace=True)

Exportando os dados para um arquivo csv

df_empresas_b3.to_csv(
    'C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\segmentacao_mercado_empresas.csv', 
    index=False)
