==> code/1_aquisicao_e_tratamento_de_dados/3_obtendo_tickers.ipynb <==
Instalando bibliotecas

# ! pip install pip install pandas
# ! pip install selenium
# ! pip install webdriver_manager

Importando Bibliotecas

import pandas as pd

Gerando lista com nomes das empresas listadas na B3

df_segm_merc_empr = pd.read_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\segmentacao_mercado_empresas.csv")
df_tickers = pd.read_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\acoes-listadas.csv")
lista_tickers = df_tickers["Código"].to_list()

df_segm_merc_empr['tickers'] = None

for ticker in lista_tickers:
    df_segm_merc_empr.loc[df_segm_merc_empr['codigo'] == ticker[0:4], 'tickers'] = ticker
    
# empresas sem tickers
df_prov = df_segm_merc_empr.loc[df_segm_merc_empr['tickers'].isna() == True]

df_prov.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\empresas_sem_negociações.csv")

# Ao pesquisar, constatou-se pelo histórico de negociações apresentados pelo site da B3
# que as sem tickers não foram negociadas até a data da realização deste trabalho (07/11/2023).
# Portanto, para evitar enviesamento das análises optou-se por removê-los.

df_segm_merc_empr = df_segm_merc_empr.loc[df_segm_merc_empr['tickers'].isna() == False]
df_segm_merc_empr.drop(['codigo'], axis=1, inplace=True)
df_segm_merc_empr.reset_index(drop=True, inplace=True)
# df_segm_merc_empr.info()
# df_segm_merc_empr.shape
display(df_segm_merc_empr)

indice_mmaq4 = df_segm_merc_empr.loc[df_segm_merc_empr["tickers"] == "MMAQ4"].index
df_segm_merc_empr = df_segm_merc_empr.drop(indice_mmaq4)

df_segm_merc_empr.loc[df_segm_merc_empr["tickers"] == "MMAQ4"].value_counts().sum()

df_segm_merc_empr.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\info_empresas.csv", index=False)

Buscando tickers das empresas listadas na B3

# lista_tickers = []

# url = "https://statusinvest.com.br/acoes"

# # abrindo o navegador
# driver = webdriver.Chrome(service=None)

# # for nome_empresa in lista_nomes_empresas:

# nome_empresa = "ENAUTA PART"

#     # acessando a url
# driver.get(url)

#     # clicando no campo de pesquisa
# WebDriverWait(driver, 5).until(
#         EC.visibility_of_element_located(
#             (By.XPATH,"/html/body/div[1]/div[3]/form/div[1]/div[1]/div[1]/div/div[2]/textarea")
#             )).click()

#     # escrevendo o ticker e o nome da empresa
# WebDriverWait(driver, 5).until(
#         EC.visibility_of_element_located(
#             (By.XPATH,"/html/body/div[1]/div[3]/form/div[1]/div[1]/div[1]/div/div[2]/textarea")
#             )).send_keys(f"ticker {nome_empresa}")

#     #
# WebDriverWait(driver, 5).until(
#         EC.visibility_of_element_located(
#             (By.XPATH,"/html/body/div[1]/div[3]/form/div[1]/div[1]/div[1]/div/div[2]/textarea")
#             )).send_keys(Keys.RETURN)


# ticker = WebDriverWait(driver, 10).until(
#         EC.visibility_of_element_located((
#             By.XPATH, "/html/body/div[5]/div/div[10]/div[2]/div/div/div[3]/div[1]/div/div/div[2]/div[2]/div[1]/div/span"))).text[-5:]
    
# # lista_tickers.append(ticker)

# print(ticker)

# # Fechando o navegador
# driver.quit()

# print(lista_tickers)

# # esperando pelo botão de aceitar coockies e clicando nele
# WebDriverWait(driver, 5).until(
#     EC.visibility_of_element_located(
#         (By.XPATH, "//button[@id='onetrust-accept-btn-handler']")
#         )).click()

# # esperando o carregamento do iframe e mudando o 
# # induza WebDriverWait para que o elemento desejado seja clicável
# WebDriverWait(driver, 5).until(
#     EC.frame_to_be_available_and_switch_to_it(
#         (By.XPATH,"//iframe[@id='bvmf_iframe']")))

# WebDriverWait(driver, 5).until(
#     EC.frame_to_be_available_and_switch_to_it(
#         (By.XPATH,"/html/body/app-root/app-companies-classification/div']")))


# # esperando pela caixa de seleção do número de itens exibidos na lista e clicando nela
# WebDriverWait(driver, 5).until(
#     EC.visibility_of_element_located(
#         (By.XPATH,"/html/body/app-root/app-companies-classification/div/div/div/app-companies-home-filter-classification/form/div[1]/div/div/select")
#         )).click()


# # # o elemento TODOS está dentro de um <iframe>, então:
# # # induza o WebDriver, espere até que o frame desejado esteja disponível e mude para ele;
# # # induza WebDriverWait para que o elemento desejado seja clicável
# # WebDriverWait(driver, 5).until(
# #     EC.frame_to_be_available_and_switch_to_it(
# #         (By.XPATH,"//iframe[@id='bvmf_iframe']")))

# # # esperando pelo botão TODOS e clicando nele
# # WebDriverWait(driver, 5).until(
# #     EC.visibility_of_element_located(
# #         (By.XPATH, "//button[@aria-label='Buscar Todas']"))).click()

# # # esperando pelo botão exibir como lista e clicando nele
# # WebDriverWait(driver, 5).until(
# #     EC.visibility_of_element_located(
# #         (By.XPATH, "/html/body/app-root/app-companies-search/div/form/div[1]/div/div/div[2]/nav/div/a[2]")
# #         )).click()

# # # esperando pela caixa de seleção do número de itens exibidos na lista e clicando nela
# # WebDriverWait(driver, 5).until(
# #     EC.visibility_of_element_located(
# #         (By.XPATH,"/html/body/app-root/app-companies-search/div/form/div[3]/div[1]/select")
# #         )).click()

# # # esperando pela abertura das opções da caixa de seleção do número 
# # # de itens exibidos na lista e clicando na opção 120
# # WebDriverWait(driver, 5).until(
# #     EC.visibility_of_element_located(
# #         (By.XPATH,"/html/body/app-root/app-companies-search/div/form/div[3]/div[1]/select/option[4]")
# #         )).click()

# # # Espere até que algum elemento esteja visível ou alguma condição seja atendida
# # empresas = WebDriverWait(driver, 5).until(
# #     EC.presence_of_element_located(
# #         (By.XPATH,'/html/body/app-root/app-companies-search/div/form/div[2]/div[2]/table'))
# #         ).text
