==> code/1_aquisicao_e_tratamento_de_dados/4_obtendo_historicos_cotacoes.ipynb <==
# ! pip install matplotlib
# ! pip install yfinance

Importando bibliotencas

import pandas as pd
import yfinance as yf

Importando dados das empresas

df_segm_merc_empr = pd.read_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\info_empresas.csv")

Obtendo Dados

# O fechamento ajustado é o preço de fechamento após os ajustes para todos os desdobramentos e distribuições de dividendos aplicáveis. Os dados são ajustados usando multiplicadores de desdobramentos e dividendos apropriados, aderindo aos Centro de pesquisa de preços de títulos (CRSP) padrões.

# https://www.crsp.org/

# Benchmark
# Índice Bovespa é o indicador do desempenho médio das cotações das ações negociadas na B3, formado pelas ações com maior volume negociado nos últimos meses.
# fonte: https://pt.wikipedia.org/wiki/Ibovespa

# Obtendo historico de variação do Ibov
hist_ibov = yf.download("^BVSP", start="2013-01-01", end="2023-10-01")["Adj Close"]
hist_ibov.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cotacoes\\hist_ibov.csv", 
                 index=True)

lista_tickers_verificacao = []

for ticker in df_segm_merc_empr["tickers"]:

    # Obtendo histórico de valores de fechamento ajustado e de volume de trasações de 2013-01-01 a 2023-10-01
    hist = yf.download(f"{ticker}.SA", start="2013-01-01", end="2023-10-01")
    hist = hist[["Adj Close", "Volume"]]

    # Validando confiabilidade dos dados obtidos
    cond_hist_obtido = hist.shape[0] > 0
    cond_sem_nulos = hist.isna().sum().iloc[0] == 0 and hist.isna().sum().iloc[1] == 0
    cond_aprov_hist = cond_hist_obtido == True and cond_sem_nulos ==True
    if cond_aprov_hist:
        # Exportando histórico validado
        hist.to_csv(f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cotacoes\\hist_{ticker}.csv",
                    index=True)
    else:
        # Adicionado ticker sem histórico validado à lista para tratamento
        lista_tickers_verificacao.append(ticker)

if len(lista_tickers_verificacao) > 0:
        # Imprimindo lista de ticker para tratamento
        print("Verificar histórico de cotações:")
        for ticker in lista_tickers_verificacao:
            print(ticker)
else:
    print("Todas os dados foram baixados e exportados com sucesso.")

hist_APTI4 = yf.download("APTI4.SA", start="2013-01-01", end="2023-10-01")
hist_APTI4 = hist_APTI4[["Adj Close", "Volume"]]

hist_CALI3 = yf.download("CALI3.SA", start="2013-01-01", end="2023-10-01")
hist_CALI3 = hist_CALI3[["Adj Close", "Volume"]]
hist_CALI3.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cotacoes\\hist_CALI3.csv", index=True)

hist_CTSA4 = yf.download("CTSA4.SA", start="2013-01-01", end="2023-10-01")
hist_CTSA4 = hist_CTSA4[["Adj Close", "Volume"]]
hist_CTSA4.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cotacoes\\hist_CTSA4.csv", index=True)

hist_PNVL3 = yf.download("PNVL3.SA", start="2013-01-01", end="2023-10-01")
hist_PNVL3 = hist_PNVL3[["Adj Close", "Volume"]]
hist_PNVL3.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cotacoes\\hist_PNVL3.csv", index=True)

hist_BPAR3 = yf.download("BPAR3.SA", start="2013-01-01", end="2023-10-01")
hist_BPAR3 = hist_BPAR3[["Adj Close", "Volume"]]

# Corrigindo tickers
df_segm_merc_empr['tickers'] = df_segm_merc_empr['tickers'].replace('CALI4', 'CALI3')
df_segm_merc_empr['tickers'] = df_segm_merc_empr['tickers'].replace('CTSA8', 'CTSA4')
df_segm_merc_empr['tickers'] = df_segm_merc_empr['tickers'].replace('PNVL4', 'PNVL3')

# Removendo tickers inexistentes
df_segm_merc_empr = df_segm_merc_empr[df_segm_merc_empr['tickers'] != "APTI4"]
df_segm_merc_empr = df_segm_merc_empr[df_segm_merc_empr['tickers'] != "BPAR3"]
df_segm_merc_empr = df_segm_merc_empr[df_segm_merc_empr['tickers'] != "MMAQ4"]
df_segm_merc_empr = df_segm_merc_empr[df_segm_merc_empr['tickers'] != "MMAQ3"]

df_segm_merc_empr.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\info_empresas.csv", index=False)
