==> code/1_aquisicao_e_tratamento_de_dados/5_consolidando_base_cotacoes.ipynb <==
Normalizando os dados

import pandas as pd
import numpy as np
import os

path_hist = 'C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cotacoes\\'
lista_historicos = os.listdir(path_hist)
# print(lista_historicos)

# Removendo tickers inexistentes
lista_tickers_removidos = ["APTI4", "BPAR3", "MMAQ4", "MMAQ3"]
lista_hist_removidos = [f"hist_{x}.csv" for x in lista_tickers_removidos]

for historico in lista_historicos:

    # print(path_hist + historico)
    if historico in lista_hist_removidos:
        os.remove(path_hist + historico)

lista_tickers_removidos = ["APTI4", "BPAR3", "MMAQ4", "MMAQ3"]
lista_hist_removidos = [f"hist_{x}.csv" for x in lista_tickers_removidos]
elementos_comuns = [elemento for elemento in lista_hist_removidos if elemento in os.listdir(path_hist)]

print(elementos_comuns)

for historico in lista_historicos:
    df = pd.read_csv(path_hist + historico, index_col="Date")
    df["Adj Close"] = df["Adj Close"]/df.iloc[0]["Adj Close"]
    df.to_csv(f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cot_norm\\{historico}", index=True)

Agrupando os ações por segmento de atuação

# Listando os segmentos de mercado
df_segm_merc_empr = pd.read_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\info_empresas.csv")

for ticker_removido in lista_tickers_removidos:
    df_segm_merc_empr = df_segm_merc_empr[df_segm_merc_empr["tickers"] != ticker_removido]
    
lista_segmentos = df_segm_merc_empr["segmento"].unique().tolist()

df_hist_cot_ticker = pd.read_csv(f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cot_norm\\hist_WEGE3.csv",
                                  index_col="Date")
df_hist_cot_ticker.index = pd.to_datetime(df_hist_cot_ticker.index)

for segmento in lista_segmentos:

    lista_tickers_segm = [x for x in df_segm_merc_empr[df_segm_merc_empr["segmento"] == segmento]["tickers"]]
    # print(lista_tickers_segm)

    data_inicial = '2013-01-01'
    data_final = '2023-10-01'
    datas = pd.date_range(start=data_inicial, end=data_final, freq='D')
    df_segm = pd.DataFrame({'Date': datas}).set_index('Date')

    for ticker in lista_tickers_segm:
        
        df_hist_cot_ticker = pd.read_csv(f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cot_norm\\hist_{ticker}.csv",
                                  index_col="Date")
        # df_hist_cot_ticker = pd.read_csv(
        #    f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cot_norm\\hist_{ticker}.csv",
        #    parse_dates= "Date",
        #    index_col="Date")
        
        for data in df_hist_cot_ticker.index:
          df_segm.loc[data, ticker] = df_hist_cot_ticker.loc[data, "Adj Close"]

    # df_segm = df_segm.dropna(how="all") # para deletar apenas as linhas vazias
    # df_segm = df_segm.dropna() # para deletar todas as linhas com algum valor vazio

    df_segm.to_csv(f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cot_norm_agrup_segm\\{segmento}.csv")

Gerando um df com o histórico de variação média de cada segmento

data_inicial = '2013-01-01'
data_final = '2023-10-01'
datas = pd.date_range(start=data_inicial, end=data_final, freq='D')
df_media_segm = pd.DataFrame({'Date': datas}).set_index('Date')

for segmento in lista_segmentos:

    df_hist_cot_ticker = pd.read_csv(
        f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cot_norm_agrup_segm\\{segmento}.csv", 
        index_col="Date")
    # df_hist_cot_ticker = pd.read_csv(
    #     f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_cot_norm_agrup_segm\\{segmento}.csv", 
    #     parse_dates= "Date",
    #     index_col="Date")
    
    for data in df_hist_cot_ticker.index:
        
        soma_linha = 0
        cont_tickers_somados = 0
        
        for ticker in df_hist_cot_ticker.columns:
            
            cond_calc_media = df_hist_cot_ticker.loc[data, ticker] !=0 and not np.isnan(df_hist_cot_ticker.loc[data, ticker])
            if cond_calc_media:
                
                soma_linha += df_hist_cot_ticker.loc[data, ticker]
                cont_tickers_somados += 1
        
        if cont_tickers_somados > 0:
            df_hist_cot_ticker.loc[data, "media"] = soma_linha / cont_tickers_somados 
    
    for data in df_hist_cot_ticker.index:
        df_media_segm.loc[data, segmento] = df_hist_cot_ticker.loc[data, "media"]

df_media_segm.head(10)

Trantando valores faltantes

for segmento in df_media_segm.columns:

    data_primeiro_ipo = df_media_segm[segmento].first_valid_index()
    # print(f"{segmento}, data_primeiro_ipo: {data_primeiro_ipo}")

    for data in df_media_segm.index:

        if data < data_primeiro_ipo:
            
            df_media_segm.loc[data, segmento] = 0

df_media_segm.drop(index="2013-01-01", inplace=True)
df_media_segm.dropna(inplace=True)
# df_media_segm.ffill(inplace=True)
df_media_segm.head(10)

df_media_segm.shape

Exportando df

df_media_segm.to_csv(f"C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\var_media_segm.csv", index=True)
