==> code/1_aquisicao_e_tratamento_de_dados/6_obtendo_hist_param_merc.ipynb <==
import pandas as pd
import yfinance as yf

Obtendo histórico de cotações Dólar/Real

usd_brl = yf.Ticker("USDBRL=X")
hist_usd_brl = pd.DataFrame(usd_brl.history(start="2013-01-01", end="2023-10-01")["Close"])

hist_usd_brl.head()

hist_usd_brl.info()

hist_usd_brl.describe()

hist_usd_brl.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_usd_brl.csv")

Obtendo histórico do IPCA

# URL do arquivo CSV do IPCA no site do Banco Central do Brasil
url = 'https://api.bcb.gov.br/dados/serie/bcdata.sgs.433/dados?formato=csv'

# Carregar o arquivo CSV no pandas
hist_ipca = pd.read_csv(url, sep=';', index_col='data', parse_dates=['data'])

# Filtrar o período desejado
hist_ipca = hist_ipca.loc['2013-01-01':'2023-10-01']

hist_ipca.head()

hist_ipca.info()

hist_ipca["valor"].str.replace(",", ".").astype(float)

hist_ipca.describe()

hist_ipca.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_ipca.csv")

Obtendo histórico do IGP-M

# URL do arquivo CSV do IGP-M no site do Banco Central do Brasil
url = 'https://api.bcb.gov.br/dados/serie/bcdata.sgs.189/dados?formato=csv'

# Carregar o arquivo CSV no pandas
hist_igpm = pd.read_csv(url, sep=';', index_col='data', parse_dates=['data'])

# Filtrar o período desejado
hist_igpm = hist_igpm.loc['2013-01-01':'2023-10-01']

hist_igpm.head()

hist_igpm.info()

hist_igpm["valor"].str.replace(",", ".").astype(float)

hist_igpm.describe()

hist_igpm.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_igpm.csv")

Obtendo historico da taxa selic

# url api: https://dadosabertos.bcb.gov.br/dataset/11-taxa-de-juros---selic/resource/3d751a0d-afb2-452b-83f2-310a201f8a82

# http://api.bcb.gov.br/dados/serie/bcdata.sgs.{11}/dados?formato={csv}&dataInicial={"01/01/2013"}&dataFinal={"01/10/2023"}

# url endpoint: https://api.bcb.gov.br/dados/serie/bcdata.sgs.11/dados?formato=csv

# documentação biblioteca sgs

# https://wilsonfreitas.github.io/python-bcb/sgs.html

# from bcb import sgs

# import matplotlib.pyplot as plt

# import matplotlib as mpl

# mpl.style.use('bmh')

# df = sgs.get({'IPCA': 433}, start='2002-02-01')

# df.index = df.index.to_period('M')

# df.head()
# Out[7]: 
#          IPCA
# Date         
# 2002-02  0.36
# 2002-03  0.60
# 2002-04  0.80
# 2002-05  0.21
# 2002-06  0.42
# dfr = df.rolling(12)

# i12 = dfr.apply(lambda x: (1 + x/100).prod() - 1).dropna() * 100

# i12.head()
# Out[10]: 
#               IPCA
# Date              
# 2003-01  14.467041
# 2003-02  15.847124
# 2003-03  16.572608
# 2003-04  16.769209
# 2003-05  17.235307
# i12.plot(figsize=(12,6))
# Out[11]: <Axes: xlabel='Date'>

# plt.title('Fonte: https://dadosabertos.bcb.gov.br', fontsize=10)
# Out[12]: Text(0.5, 1.0, 'Fonte: https://dadosabertos.bcb.gov.br')

# plt.suptitle('IPCA acumulado 12 meses - Janela Móvel', fontsize=18)
# Out[13]: Text(0.5, 0.98, 'IPCA acumulado 12 meses - Janela Móvel')

# plt.xlabel('Data')
# Out[14]: Text(0.5, 0, 'Data')

# plt.ylabel('%')
# Out[15]: Text(0, 0.5, '%')

# plt.legend().set_visible(False)

import pandas as pd

# Faz a requisição HTTP para a API do Banco Central
url = 'https://api.bcb.gov.br/dados/serie/bcdata.sgs.11/dados?formato=json'
hist_selic = pd.read_json(url)
hist_selic["data"] = pd.to_datetime(hist_selic["data"], format="%d/%m/%Y")
hist_selic.set_index('data', inplace=True)
hist_selic = hist_selic.loc["2012-01-01":"2023-10-01"]
hist_selic["ano_mes"] = hist_selic.index.strftime('%Y-%m')
hist_selic.drop_duplicates()
# hist_selic_agrupado = hist_selic.groupby(hist_selic.index.month)
# hist_selic_agrupado = hist_selic_agrupado["valor"].mean()
#hist_selic["valor"] = hist_selic["valor"].rolling(12).apply(lambda x: (1 + x/100).prod() - 1).dropna() * 100
display(hist_selic["ano_mes"])
# hist_selic.tail(20)

hist_selic.isna().sum()

# display(hist_selic)

! pip install sgs

fromimport sgs

selic = sgs.get({"selic":432}, start="2013-01-01", end="2023-10-01")

selic.tail()

# selic.head()

# selic.info()

# selic.describe()

# selic.to_csv("C:\\Users\\victo\\OneDrive\\Documentos\\github\\tcc-analise-setores-mercado-financeiro\\data\\hist_selic.csv")
